Refactor to use an enum for error codes
