+++ src/models/requirements/number/numberIntegerRequirement.ts
@@ -1,14 +1,14 @@
-import type { RequirementValidationError } from '../../../types/ValidationResult'
+import { RequirementErrorCodes, type RequirementValidationError } from '../../../types/ValidationResult'
 import Requirement from '../reqirement'
 
 export default class NumberIntegerRequirement extends Requirement {
   public validate(value: number): RequirementValidationError[] {
-    // If the remainder after dividing by one is zero, it is an integer
+    // Any remainder after dividing by 1 means that the number is not an integer.
     if (value % 1 === 0) return []
 
     return [
       {
-        code: 'pattern',
+        code: RequirementErrorCodes.PATTERN,
         message: `Value of ${value} is not an integer.`
       }
     ]
