+++ src/models/requirements/string/stringRequirement.ts
@@ -1,4 +1,4 @@
-import type { RequirementValidationError } from '../../../types/ValidationResult'
+import { RequirementErrorCodes, type RequirementValidationError } from '../../../types/ValidationResult'
 import Requirement from '../reqirement'
 
 export default class StringRequirement extends Requirement {
@@ -7,7 +7,7 @@ export default class StringRequirement extends Requirement {
 
     return [
       {
-        code: 'type',
+        code: RequirementErrorCodes.TYPE,
         message: 'Value is not a valid string.'
       }
     ]
