+++ src/types/VariableTypes.d.ts
@@ -1,0 +1,3 @@
+type VariableTypes = 'string' | 'bigint' | 'number' | 'boolean' | 'symbol' | 'undefined' | 'object' | 'function'
+
+export default VariableTypes
